pyproject.toml
[tool.poetry]
name = "kodiak"
version = "0.1.0"
description = "Kodiak - LLM Penetration Testing Suite with TUI"
authors = ["Kodiak Team"]
readme = "README.md"
packages = [{include = "kodiak"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
pydantic = {extras = ["email"], version = "^2.6.0"}
pydantic-settings = "^2.1.0"
litellm = "^1.20.0"
sqlalchemy = "^2.0.25"
sqlmodel = "^0.0.14"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
psycopg = {extras = ["binary"], version = "^3.1.18"}
python-multipart = "^0.0.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
websockets = "^12.0"
docker = "^7.0.0"
jinja2 = "^3.1.3"
httpx = "^0.26.0"
tenacity = "^8.2.3"
sentry-sdk = "^1.40.0"
loguru = "^0.7.2"
playwright = "^1.40.0"
pyyaml = "^6.0.1"
textual = "^0.48.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
hypothesis = "^6.98.0"
black = "^24.1.0"
isort = "^5.13.0"
mypy = "^1.8.0"
ruff = "^0.1.15"

[tool.poetry.scripts]
kodiak = "kodiak.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.black]
line-length = 120
target-version = ['py311']
